pyproject.toml
[tool.poetry]
name = "softphone-airflow-provider"
version = "0.1.0"
description = ""
authors = ["Rostislav Aleev <rs.aleev@gmail.com>"]
readme = "README.md"
packages = [{include = "softphone_airflow_provider"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.30.0"


[tool.poetry.group.dev.dependencies]
apache-airflow = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
